[Fix] Mirror URL format
Problem:
Puppet agent isn't running, thus causing the dependent
services to not function properly.

Cause:
Because of the trailing slash in the mirror_url config,
reposiroty_mirror_sync was parsing the source_path including
the trailing slash, and another "/" was getting appended to it
to form the URL for wget.

Solution:
Removing the trailing slash from the mirror_url will fix the problem.

Resolves: https://gitea.obmondo.com/EnableIT/abbnoa6nlk/issues/857, https://gitea.obmondo.com/EnableIT/abbnoa6nlk/issues/858, https://gitea.obmondo.com/EnableIT/abbnoa6nlk/issues/859, and https://gitea.obmondo.com/EnableIT/abbnoa6nlk/issues/860
Assisted-by: Kanha Gupta <kanha@obmondo.com>
Signed-off-by: Sidharth Jawale <sidharth@obmondo.com>