feat: written a custom facter, to get the certname directly from the certname.
The issue is chicken and egg, since the puppet_settings facter does not resilve, due to ttls introduce in the facter.conf

The above facter does not resolve and obmondo_customer fails to load, since its dependent on the puppet_settings.

I tried to load a custom facter, but since the puppet catalog is failing to compile, so the file resource would never run, maybe we can do with tags, but doing this on all is not possible.

The new facter can be loaded before creating the catalog, which will give the required value for obmondo_customer facter